+++ src/main/java/lotto/application/WinningStatistics.java
@@ -6,7 +6,7 @@ import lotto.domain.Lotto;
 import lotto.domain.LottoCount;
 import lotto.domain.LottoRank;
 import lotto.domain.Lottos;
-import lotto.dto.WinningLottoInformation;
+import lotto.domain.WinningLottoInformation;
 import lotto.dto.request.LottoCalculationRequest;
 import lotto.dto.response.LottoCalculationResponse;
