+++ src/test/java/lotto/application/WinningStatisticsTest.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import lotto.domain.Lotto;
 import lotto.domain.LottoCount;
 import lotto.domain.Lottos;
-import lotto.dto.WinningLottoInformation;
+import lotto.domain.WinningLottoInformation;
 import lotto.dto.request.LottoCalculationRequest;
 import lotto.dto.response.LottoCalculationResponse;
 import org.assertj.core.api.Assertions;
